requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pydantic==1.10.17
prometheus_client==0.19.0
httpx==0.27.0
pytest==8.3.2
pytest-cov==5.0.0
langdetect==1.0.9
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
openai==1.35.14
requests>=2.31.0
pytest-cov>=4.1.0
